+++ resources/views/application.blade.php
@@ -10,8 +10,6 @@
     <title>{{ config('app.name', 'PILOS') }}</title>
 
     <!-- Scripts -->
-    <script src="{{ asset('js/manifest.js') }}" defer></script>
-    <script src="{{ asset('js/vendor.js') }}" defer></script>
     <script src="{{ asset('js/app.js') }}" defer></script>